Move the backend mount code into separate modules too.
