+++ src/backend/linux_raw/mod.rs
@@ -51,6 +51,10 @@ pub(crate) mod io;
 pub(crate) mod io_uring;
 #[cfg(feature = "mm")]
 pub(crate) mod mm;
+#[cfg(feature = "mount")]
+pub(crate) mod mount;
+#[cfg(all(feature = "fs", not(feature = "mount")))]
+pub(crate) mod mount; // for deprecated mount functions in "fs"
 #[cfg(feature = "net")]
 pub(crate) mod net;
 #[cfg(any(
