+++ src/mount/mod.rs
@@ -2,14 +2,18 @@
 
 // The `mount` module includes the `mount` function and related
 // functions which were originally defined in `rustix::fs` but are
-// now replaced by deprecated aliases.
+// now replaced by deprecated aliases. After the next semver bump,
+// we can remove the aliases and all the `#[cfg(feature = "mount")]`
+// here and in src/backend/*/mount.
 //
 // The `fsopen` module includes `fsopen` and related functions.
 
 #[cfg(feature = "mount")]
 mod fsopen;
 mod mount;
+mod types;
 
 #[cfg(feature = "mount")]
 pub use fsopen::*;
 pub use mount::*;
+pub use types::*;
